Refactor utils and delta_loss modules to enhance logging and expert score handling; streamline quantization scheme representation and layer configuration logging.
