+++ auto_round/auto_scheme/utils.py
@@ -11,26 +11,33 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import logging
+import math
 import re
 from dataclasses import asdict, fields
-from typing import Iterable, Union
+from typing import Iterable, Optional, Union
 
 import torch
 from accelerate import dispatch_model, infer_auto_device_map
 from accelerate.utils import get_balanced_memory
 
-from auto_round.schemes import QuantizationScheme, preset_name_to_scheme
+from auto_round.logger import logger
+from auto_round.schemes import QuantizationScheme, preset_name_to_scheme, scheme_to_preset_name
 from auto_round.utils import (
     DEVICE_ENVIRON_VARIABLE_MAPPING,
     SUPPORTED_LAYER_TYPES,
     check_to_quantized,
     get_block_names,
     get_layer_features,
+    get_lm_head_name,
     get_module,
     is_hpex_available,
     normalize_no_split_modules,
     parse_available_devices,
 )
+
+_EXPERT_ID_PATTERN = re.compile(r"^(.*?\.experts\.\d+)(?:\.|$)")
+_ZERO_EPS = 1e-12
 
 
 def apply_quant_scheme(
@@ -332,3 +339,423 @@ def parse_shared_layers(model: torch.nn.Module, shared_patterns: Iterable[Iterab
                 matched_groups.append(sorted(matched_layers))
     matched_groups = merge_lists_unionfind(matched_groups)
     return matched_groups
+
+
+def _expert_key_from_layer_name(layer_name: str) -> Optional[str]:
+    """Map one MoE-related linear layer to a unique expert key.
+
+    Gate/up/down projections belonging to the same expert should map to one key.
+    """
+    match = _EXPERT_ID_PATTERN.search(layer_name)
+    if match:
+        return match.group(1)
+
+    if ".moe." in layer_name:
+        # Fallback for models that expose MoE expert layers without
+        # `.experts.<id>` naming. Best-effort collapse by dropping the last
+        # projection suffix (gate/up/down/etc.).
+        parts = layer_name.split(".")
+        if len(parts) > 1:
+            return ".".join(parts[:-1])
+        return layer_name
+
+    return None
+
+
+def _short_summary_name(layer_name: str) -> str:
+    """Shorten ``layer_name`` to its last two dotted segments when the final segment is a
+    numeric index (e.g. expert id), otherwise return it unchanged.
+    """
+    parts = layer_name.rsplit(".", 2)
+    if len(parts) >= 2 and parts[-1].isdigit():
+        return ".".join(parts[-2:])
+    return layer_name
+
+
+def _scheme_short_name(scheme) -> str:
+    """Concise human-readable label for a quantization scheme, e.g. 'MXFP4' or 'W4A16'."""
+    if isinstance(scheme, str):
+        return scheme
+    if isinstance(scheme, QuantizationScheme):
+        preset_name = scheme_to_preset_name(scheme)
+        if preset_name:
+            return preset_name
+        scheme = asdict(scheme)
+    if isinstance(scheme, dict):
+        try:
+            preset_name = scheme_to_preset_name(QuantizationScheme.from_dict(scheme))
+            if preset_name:
+                return preset_name
+        except Exception:  # noqa: BLE001
+            pass
+        data_type = str(scheme.get("data_type", "")).lower()
+        bits = scheme.get("bits", 16)
+        act_bits = scheme.get("act_bits", 16)
+        if data_type.startswith("mx_fp"):
+            return f"MXFP{bits}"
+        if data_type.startswith("nv_fp"):
+            return f"NVFP{bits}"
+        if data_type.startswith("gguf"):
+            return data_type.upper()
+        if act_bits >= 16:
+            return f"W{bits}"
+        return f"W{bits}A{act_bits}"
+    return str(scheme)
+
+
+def _format_scientific_loss(value: float) -> str:
+    """Format a score using compact scientific notation like 1.23E-5."""
+    text = f"{value:.2E}"
+    return re.sub(r"E([+-])0+(\d+)", r"E\1\2", text)
+
+
+def _fill_inactive_expert_scores(scores_dict: dict[str, list[float]], block_names: list[str]):
+    """Fill inactive experts with the average loss of active experts in each block.
+
+    Inactive expert means all its tracked projection losses are zero.
+    """
+    block_prefixes = [(name, name + ".") for name in sorted(block_names, key=len, reverse=True)]
+    block_expert_stats: dict[str, dict[str, dict[str, object]]] = {}
+
+    for layer_name, values in scores_dict.items():
+        layer_loss = float(values[1])
+        if not math.isfinite(layer_loss):
+            continue
+
+        matched_block = None
+        for block_name, block_prefix in block_prefixes:
+            if layer_name == block_name or layer_name.startswith(block_prefix):
+                matched_block = block_name
+                break
+        if matched_block is None:
+            continue
+
+        expert_key = _expert_key_from_layer_name(layer_name)
+        if expert_key is None:
+            continue
+
+        expert_stats = block_expert_stats.setdefault(matched_block, {}).setdefault(
+            expert_key,
+            {"layers": [], "has_active": False},
+        )
+        expert_stats["layers"].append(layer_name)
+
+        if abs(layer_loss) >= _ZERO_EPS:
+            expert_stats["has_active"] = True
+
+    for block_name, expert_stats_map in block_expert_stats.items():
+        active_expert_avg_losses = []
+        for expert_stats in expert_stats_map.values():
+            if not expert_stats["has_active"]:
+                continue
+            active_losses = [
+                float(scores_dict[layer_name][1])
+                for layer_name in expert_stats["layers"]
+                if abs(float(scores_dict[layer_name][1])) >= _ZERO_EPS
+            ]
+            if not active_losses:
+                continue
+            active_expert_avg_losses.append(sum(active_losses) / len(active_losses))
+        if not active_expert_avg_losses:
+            continue
+        fill_value = sum(active_expert_avg_losses) / len(active_expert_avg_losses)
+        for _, expert_stats in expert_stats_map.items():
+            if expert_stats["has_active"]:
+                continue
+            for layer_name in expert_stats["layers"]:
+                scores_dict[layer_name][1] = fill_value
+
+
+def _log_score_summary_by_block_and_nonblock(
+    scores_dict: dict[str, list[float]],
+    block_names: list[str],
+    model=None,
+    scheme_tag: Optional[str] = None,
+    summary_stage: Optional[str] = None,
+):
+    """Log a per-block (and non-block) breakdown of ``scores_dict`` losses at debug level.
+
+    For MoE blocks, additionally splits each block's loss into non-expert vs. expert
+    contributions and reports how many experts were never activated.
+    """
+    if not scores_dict:
+        logger.info("AutoScheme score summary: empty.")
+        return
+
+    head_name = get_lm_head_name(model) if model is not None else None
+    if head_name is not None and head_name.endswith(".orig_layer"):
+        head_name = head_name[: -len(".orig_layer")]
+    if head_name is None and "lm_head" in scores_dict:
+        head_name = "lm_head"
+
+    block_prefixes = [(name, name + ".") for name in sorted(block_names, key=len, reverse=True)]
+    block_stats: dict[str, list[float]] = {name: [0.0, 0.0] for name in block_names}
+    block_expert_seen: dict[str, set[str]] = {name: set() for name in block_names}
+    block_expert_active: dict[str, set[str]] = {name: set() for name in block_names}
+    block_expert_losses: dict[str, dict[str, list[float]]] = {}
+    block_regular_losses: dict[str, list[float]] = {name: [] for name in block_names}
+    non_block_items: list[tuple[str, float]] = []
+
+    for layer_name, values in scores_dict.items():
+        layer_loss = float(values[1])
+        if not math.isfinite(layer_loss):
+            continue
+        matched_block = None
+        for block_name, block_prefix in block_prefixes:
+            if layer_name == block_name or layer_name.startswith(block_prefix):
+                matched_block = block_name
+                break
+        if matched_block is None:
+            if layer_name != head_name:
+                non_block_items.append((layer_name, layer_loss))
+        else:
+            expert_key = _expert_key_from_layer_name(layer_name)
+            if expert_key is None:
+                block_regular_losses[matched_block].append(layer_loss)
+            else:
+                block_expert_losses.setdefault(matched_block, {}).setdefault(expert_key, []).append(layer_loss)
+
+        expert_key = _expert_key_from_layer_name(layer_name)
+        if expert_key is not None and matched_block is not None:
+            block_expert_seen[matched_block].add(expert_key)
+            if abs(layer_loss) >= _ZERO_EPS:
+                block_expert_active[matched_block].add(expert_key)
+
+    for block_name in block_names:
+        for layer_loss in block_regular_losses.get(block_name, []):
+            block_stats[block_name][0] += layer_loss
+            block_stats[block_name][1] += 1
+        for losses in block_expert_losses.get(block_name, {}).values():
+            if not any(abs(loss) >= _ZERO_EPS for loss in losses):
+                continue
+            for layer_loss in losses:
+                block_stats[block_name][0] += layer_loss
+                block_stats[block_name][1] += 1
+
+    has_moe_block = any(block_expert_seen.get(block_name) for block_name in block_names)
+    tag = f"[{scheme_tag}] " if scheme_tag else ""
+    stage_str = f" ({summary_stage})" if summary_stage else ""
+    logger.debug("AutoScheme %sblock loss summary%s:", tag, stage_str)
+    if has_moe_block:
+        logger.debug("AutoScheme | block | avg_loss | non_exp_avg | exp_avg | inactive_exp | shared_loss |")
+    else:
+        logger.debug("AutoScheme | block | avg_loss |")
+
+    for block_name in block_names:
+        total_loss, cnt = block_stats.get(block_name, [0.0, 0.0])
+        avg_loss = 0.0 if cnt <= 0 else total_loss / cnt
+        expert_total = len(block_expert_seen.get(block_name, set()))
+        expert_inactive = expert_total - len(block_expert_active.get(block_name, set()))
+        regular_losses = block_regular_losses.get(block_name, [])
+        non_expert_avg_loss = None
+        if expert_total and regular_losses:
+            non_expert_avg_loss = sum(regular_losses) / len(regular_losses)
+        expert_active_losses = []
+        for losses in block_expert_losses.get(block_name, {}).values():
+            if not any(abs(loss) >= _ZERO_EPS for loss in losses):
+                continue
+            expert_active_losses.extend(losses)
+        expert_avg_loss = None
+        if expert_total and expert_active_losses:
+            expert_avg_loss = sum(expert_active_losses) / len(expert_active_losses)
+        shared_loss = None
+        if expert_inactive:
+            active_expert_avg_losses = []
+            for losses in block_expert_losses.get(block_name, {}).values():
+                active_losses = [loss for loss in losses if abs(loss) >= _ZERO_EPS]
+                if not active_losses:
+                    continue
+                active_expert_avg_losses.append(sum(active_losses) / len(active_losses))
+            if active_expert_avg_losses:
+                shared_loss = sum(active_expert_avg_losses) / len(active_expert_avg_losses)
+
+        if has_moe_block:
+            non_expert_avg = "N/A" if non_expert_avg_loss is None else f"{non_expert_avg_loss:.6f}"
+            expert_avg = "N/A" if expert_avg_loss is None else f"{expert_avg_loss:.6f}"
+            shared_avg = "N/A" if shared_loss is None else f"{shared_loss:.6f}"
+            logger.debug(
+                "AutoScheme | %s | %.6f | %s | %s | %d/%d | %s |",
+                _short_summary_name(block_name),
+                avg_loss,
+                non_expert_avg,
+                expert_avg,
+                expert_inactive,
+                expert_total,
+                shared_avg,
+            )
+        else:
+            logger.debug("AutoScheme | %s | %.6f |", _short_summary_name(block_name), avg_loss)
+
+    if head_name is not None:
+        head_loss = None
+        if head_name in scores_dict:
+            head_loss = float(scores_dict[head_name][1])
+        head_avg = "N/A" if head_loss is None or not math.isfinite(head_loss) else f"{head_loss:.6f}"
+        if has_moe_block:
+            logger.debug("AutoScheme | %s | %s | N/A | N/A | 0/0 | N/A |", head_name, head_avg)
+        else:
+            logger.debug("AutoScheme | %s | %s |", head_name, head_avg)
+
+    if has_moe_block:
+        logger.debug(
+            "AutoScheme table note: non_exp_avg excludes experts; exp_avg excludes inactive experts; "
+            "shared_loss is used for inactive expert broadcast."
+        )
+
+    if non_block_items:
+        non_block_items.sort(key=lambda x: x[0])
+        for layer_name, layer_loss in non_block_items:
+            logger.info("AutoScheme non_block=%s loss=%.6f", layer_name, layer_loss)
+    else:
+        logger.info("AutoScheme non_block loss: none")
+
+
+def _collect_current_scores(model):
+    """Snapshot the current ``mix_score`` of every wrapped module into a
+    ``{name: [0.0, loss]}`` dict.
+    """
+    scores_dict = {}
+    for name, module in model.named_modules():
+        if not hasattr(module, "mix_score"):
+            continue
+        loss = float(module.mix_score)
+        if not math.isfinite(loss):
+            continue
+        scores_dict[name] = [0.0, loss]
+    return scores_dict
+
+
+_BATCH_SUMMARY_LOG_INTERVAL = 10
+
+
+def _log_batch_avg_loss(model, batch_idx: int, pbar=None, block_names=None, total_batches=None, scheme_tag=None):
+    """Log the running average ``mix_score`` after processing one calibration batch."""
+    is_last_batch = total_batches is not None and batch_idx == total_batches
+    should_log = is_last_batch or batch_idx % _BATCH_SUMMARY_LOG_INTERVAL == 0 or logger.isEnabledFor(logging.DEBUG)
+    if not should_log:
+        return
+
+    total_loss = 0.0
+    layer_cnt = 0
+    for _, module in model.named_modules():
+        if not hasattr(module, "mix_score"):
+            continue
+        loss = float(module.mix_score)
+        if not math.isfinite(loss):
+            continue
+        total_loss += loss
+        layer_cnt += 1
+
+    avg_loss = 0.0 if layer_cnt == 0 else total_loss / layer_cnt
+    tag = f"[{scheme_tag}] " if scheme_tag else ""
+    batch_str = f"{batch_idx}/{total_batches}" if total_batches is not None else str(batch_idx)
+    msg = f"AutoScheme {tag}cumulative batch {batch_str}  avg_loss={avg_loss:.6f} layers={layer_cnt}"
+    if pbar is not None:
+        pbar.write(msg)
+    logger.debug(msg)
+
+    if block_names:
+        scores_dict = _collect_current_scores(model)
+        if scores_dict:
+            tag = f"[{scheme_tag}] " if scheme_tag else ""
+            batch_str = f"{batch_idx}/{total_batches}" if total_batches is not None else str(batch_idx)
+            if is_last_batch:
+                logger.debug("AutoScheme %scumulative batch %s block summary skipped (same as final)", tag, batch_str)
+            else:
+                logger.debug("AutoScheme %scumulative batch %s block summary:", tag, batch_str)
+                _log_score_summary_by_block_and_nonblock(
+                    scores_dict,
+                    block_names,
+                    model=model,
+                    scheme_tag=scheme_tag,
+                    summary_stage="cumulative",
+                )
+
+
+def _describe_layer_config(layer_config, scores_dict, block_names, model=None):
+    """Log a compact, structured view of the final ``layer_config``."""
+
+    if not layer_config:
+        logger.info("AutoScheme final layer_config: empty.")
+        return
+
+    head_name = get_lm_head_name(model) if model is not None else None
+    if head_name is not None and head_name.endswith(".orig_layer"):
+        head_name = head_name[: -len(".orig_layer")]
+    block_prefixes = [(name, name + ".") for name in sorted(block_names, key=len, reverse=True)]
+
+    block_rows: dict[str, list[tuple[str, str, str]]] = {name: [] for name in block_names}
+    block_expert_scheme_counts: dict[str, dict[str, int]] = {name: {} for name in block_names}
+    block_expert_layers: dict[str, dict[str, list[tuple[str, str]]]] = {name: {} for name in block_names}
+    head_row: tuple[str, str] | None = None
+    other_rows: list[tuple[str, str, str]] = []
+
+    for layer_name in sorted(layer_config):
+        scheme_name = _scheme_short_name(layer_config[layer_name])
+        mix_score = scores_dict.get(layer_name, [None, None])[1]
+        mix_score_str = (
+            None
+            if mix_score is None or not math.isfinite(float(mix_score))
+            else _format_scientific_loss(float(mix_score))
+        )
+
+        if head_name is not None and layer_name == head_name:
+            head_row = (layer_name, scheme_name)
+            continue
+
+        matched_block = None
+        for block_name, block_prefix in block_prefixes:
+            if layer_name == block_name or layer_name.startswith(block_prefix):
+                matched_block = block_name
+                break
+
+        expert_key = _expert_key_from_layer_name(layer_name)
+        if matched_block is None:
+            other_rows.append((layer_name, scheme_name, mix_score_str or "N/A"))
+            continue
+
+        if expert_key is not None:
+            block_expert_layers[matched_block].setdefault(expert_key, []).append((layer_name, scheme_name))
+            continue
+
+        short_name = layer_name[len(matched_block) + 1 :] if layer_name.startswith(matched_block + ".") else layer_name
+        block_rows[matched_block].append((short_name, scheme_name, mix_score_str or "N/A"))
+
+    for block_name, expert_map in block_expert_layers.items():
+        if not expert_map:
+            continue
+        counts: dict[str, int] = {}
+        for expert_layers in expert_map.values():
+            scheme_set = {scheme_name for _, scheme_name in expert_layers}
+            rep_scheme = (
+                next(iter(sorted(scheme_set))) if len(scheme_set) == 1 else f"MIXED({', '.join(sorted(scheme_set))})"
+            )
+            counts[rep_scheme] = counts.get(rep_scheme, 0) + 1
+        block_expert_scheme_counts[block_name] = counts
+
+    logger.info("AutoScheme final layer_config:")
+    for block_name in block_names:
+        rows = block_rows.get(block_name, [])
+        expert_counts = block_expert_scheme_counts.get(block_name, {})
+        if not rows and not expert_counts:
+            continue
+
+        logger.info("  %s", block_name)
+        for short_name, scheme_name, mix_score_str in sorted(rows, key=lambda item: item[0]):
+            logger.info("    %-48s %-10s mix_score=%s", short_name, scheme_name, mix_score_str)
+
+        if expert_counts:
+            expert_total = len(block_expert_layers.get(block_name, {}))
+            expert_summary = ", ".join(
+                f"{count}/{expert_total} {scheme_name}"
+                for scheme_name, count in sorted(expert_counts.items(), key=lambda item: (-item[1], item[0]))
+            )
+            logger.info("    experts: %s", expert_summary)
+
+    if other_rows:
+        logger.info("  other")
+        for layer_name, scheme_name, mix_score_str in other_rows:
+            logger.info("    %-48s %-10s mix_score=%s", layer_name, scheme_name, mix_score_str)
+
+    if head_row is not None:
+        logger.info("  %s is using %s", head_row[0], head_row[1])
